backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
# readme = "README.md" # Intentionally kept commented out from previous fix
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    # Add other classifiers as appropriate
]

dependencies = [
    "pydantic>=2.11.4",
    "fastapi[standard]>=0.115.12",
    "psycopg[binary]>=3.2.9",
    "pyjwt>=2.10.1",
    "bcrypt>=4.3.0",
    "transformers",
    "torch",
    "bitsandbytes",
    "accelerate",
    "fastembed",
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]